Move ASTAccessor to actions module.
